requirements.txt
alembic==1.8.1
amqp==5.1.1
async-timeout==4.0.2
attrs==22.1.0
autopage==0.5.1
azure-core==1.26.0
azure-identity==1.11.0
azure-storage-blob==12.14.1
backoff==2.2.1
billiard==3.6.4.0
black==22.12.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cliff==4.0.0
cmaes==0.9.0
cmd2==2.4.2
colorlog==6.7.0
contourpy==1.0.6
cryptography==38.0.1
cycler==0.11.0
DateTime==4.7
Deprecated==1.2.13
dnspython==2.3.0
ElasticEmail==4.0.20
et-xmlfile==1.1.0
Flask==2.2.2
Flask-Cors==3.0.10
flask-extended==0.2
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.5.1
flower==1.2.0
fonttools==4.38.0
greenlet==1.1.3.post0
humanize==4.4.0
idna==3.4
importlib-metadata==5.0.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
kombu==5.2.4
mailchimp-marketing==3.0.80
Mako==1.2.3
MarkupSafe==2.1.1
matplotlib==3.5.0
monotonic==1.6
msal==1.20.0
msal-extensions==1.0.0
msrest==0.7.1
mypy-extensions==0.4.3
numpy==1.23.0
oauthlib==3.2.2
openpyxl==3.0.10
optuna==2.10.0
packaging==21.3
pandas==1.4.0
pathspec==0.10.3
patsy==0.5.3
pbr==5.11.0
Pillow==9.3.0
platformdirs==2.6.2
plotly==5.11.0
portalocker==2.6.0
prettytable==3.5.0
prometheus-client==0.15.0
prompt-toolkit==3.0.31
psycopg-binary==3.1.4
psycopg2-binary==2.9.5
pyaml==21.10.1
pycparser==2.21
PyJWT==2.6.0
pymongo==4.3.3
pyparsing==3.0.9
pyperclip==1.8.2
python-dateutil==2.8.2
pytz==2022.5
PyYAML==6.0
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
scikit-learn==1.1.0
scikit-optimize==0.9.0
scipy==1.9.3
seaborn==0.11.0
segment-analytics-python==2.2.2
setuptools-scm==7.0.5
six==1.16.0
skforecast==0.5.1
SQLAlchemy==1.4.41
statsmodels==0.13.0
stevedore==4.1.0
tenacity==8.1.0
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.2
tqdm==4.64.0
typing_extensions==4.4.0
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
xgboost==1.6.2
zipp==3.10.0
zope.interface==5.5.1
